Splitting signup components
