+++ src/components/auth/Inputfield.tsx
@@ -1,0 +1,35 @@
+type InputFieldProps = {
+  label: string;
+  name: string;
+  type?: string;
+  value: string;
+  required?: boolean;
+  placeholder?: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+};
+
+export const InputField = ({
+  label,
+  name,
+  type = 'text',
+  value,
+  required = false,
+  placeholder,
+  onChange,
+}: InputFieldProps): React.JSX.Element => (
+  <div>
+    <label className="block text-sm font-medium text-gray-700 mb-1">
+      {label} {required && <span className="text-red-500">*</span>}
+    </label>
+    <input
+      type={type}
+      name={name}
+      value={value}
+      onChange={onChange}
+      required={required}
+      placeholder={placeholder}
+      className="w-full px-4 py-2 border border-gray-300 rounded-lg 
+                 focus:ring-2 focus:ring-blue-500 focus:outline-none"
+    />
+  </div>
+);
