+++ src/type/auth.types.ts
@@ -1,0 +1,6 @@
+export type SignUpFormDataType = {
+  fullname: string;
+  email: string;
+  password: string;
+  profileImage: File | null;
+};
